+++ extensions/dear-implot/tests/thread_safety.rs
@@ -14,6 +14,7 @@ fn thread_safety_implot_scope_tokens_not_send_sync() {
     assert_not_impl_any!(dear_implot::LegendToken: Send, Sync);
     assert_not_impl_any!(dear_implot::StyleVarToken: Send, Sync);
     assert_not_impl_any!(dear_implot::StyleColorToken: Send, Sync);
+    assert_not_impl_any!(dear_implot::ColormapToken: Send, Sync);
     assert_not_impl_any!(dear_implot::AxisFormatterToken: Send, Sync);
     assert_not_impl_any!(dear_implot::AxisTransformToken: Send, Sync);
 }